Custom Exception implemented

+++ src/main/java/com/livares/intern/product/services/impl/CategoryServiceImpl.java
@@ -4,9 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.web.util.ThrowableCauseExtractor;
 import org.springframework.stereotype.Service;
 
 import com.livares.intern.product.dto.categoryDTO;
+import com.livares.intern.product.exception.CustomException;
+import com.livares.intern.product.exception.CustomExceptionHandler;
+import com.livares.intern.product.exception.ErrorCode;
 import com.livares.intern.product.models.Category;
 import com.livares.intern.product.repository.CategoryRepository;
 import com.livares.intern.product.services.CategoryServices;
@@ -26,7 +30,11 @@ public class CategoryServiceImpl implements CategoryServices {
 	//retrieving specific category data
 	@Override
 	public List<categoryDTO> getCategoryById(Long id){
+		if(categoryRepository.existsById(id))
 			return categoryRepository.findCategoryById(id);
+		else {
+			throw new CustomException(ErrorCode.NOT_FOUND,"Could not find the category id");
+		}
 		}
 	
 	//creating a new category field
@@ -40,10 +48,14 @@ public class CategoryServiceImpl implements CategoryServices {
 	//deleting  a category
 	@Override
 	public void deleteCategoryById(Long id) {
+		if(categoryRepository.existsById(id))
 		 categoryRepository.deleteById(id);
+		else {
+			throw new CustomException(ErrorCode.NOT_FOUND,"Could not find the category id : "+id);
+		}
 	}
 
-	
+	//updating a category
 	@Override
 	public Category updateCategory(Long id, categoryDTO category) {
 		Optional <Category> optionalCategory = categoryRepository.findById(id);
@@ -53,23 +65,10 @@ public class CategoryServiceImpl implements CategoryServices {
 			return categoryRepository.save(existingCategory);
 		}
 		else {
-			throw new RuntimeException("Category not found with id: " + id);
+			throw new CustomException(ErrorCode.NOT_FOUND,"Could not find the category id: "+id);//using custom exception
+			
+			//throw new RuntimeException("Category not found with id: " + id);
 		}
 	}
 	
-	//updating category filed
-//	public Category updateCategory(Long id,categoryDTO category) {
-//		Optional <Category> optionalCategory =categoryRepository.findById(id);
-//		if(optionalCategory.isPresent()) {
-//			Category existingCategory = optionalCategory.get();
-////			existingCategory.setId(category.getId());
-//			existingCategory.setCategory(category.getCategory());
-//			return categoryRepository.save(existingCategory);
-//		}
-//		else {
-//			throw new RuntimeException("Category not found with id: " + id);
-//		}
-//		
-//	}
-
 }
